# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: Lumière Beauty API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /products:
    get:
      operationId: getProducts
      tags: [products]
      summary: List products with filters
      parameters:
        - name: category
          in: query
          schema:
            type: string
        - name: subcategory
          in: query
          schema:
            type: string
        - name: skinType
          in: query
          schema:
            type: string
        - name: minPrice
          in: query
          schema:
            type: number
        - name: maxPrice
          in: query
          schema:
            type: number
        - name: minRating
          in: query
          schema:
            type: number
        - name: inStock
          in: query
          schema:
            type: boolean
        - name: sort
          in: query
          schema:
            type: string
            enum: [bestSelling, newArrivals, priceLow, priceHigh, topRated]
        - name: search
          in: query
          schema:
            type: string
        - name: page
          in: query
          schema:
            type: integer
        - name: limit
          in: query
          schema:
            type: integer
      responses:
        "200":
          description: Products list
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ProductsResponse"
...
  /products/{slug}:
    get:
      operationId: getProductBySlug
      tags: [products]
      summary: Get product by slug
      parameters:
        - name: slug
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Product detail
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ProductDetail"
  /products/{id}/related:
    get:
      operationId: getRelatedProducts
      tags: [products]
      summary: Get related products
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Related products
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Product"

  # ADMIN PRODUCTS
...
  /reviews/{productId}:
    get:
      operationId: getReviews
      tags: [reviews]
      summary: Get reviews for a product
      parameters:
        - name: productId
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Reviews
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Review"
    post:
      operationId: createReview
      tags: [reviews]
      summary: Create a review
      parameters:
        - name: productId
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateReviewBody"
      responses:
        "201":
          description: Review created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Review"

  # QUIZ
  /quiz/submit:
    post:
      operationId: submitQuiz
      tags: [quiz]
      summary: Submit quiz and get AI recommendations
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/QuizAnswers"
      responses:
        "200":
          description: Quiz results with recommendations
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/QuizResult"
...
components:
  schemas:
...
    Product:
      type: object
      properties:
        id:
          type: integer
        name:
          type: string
        slug:
          type: string
        category:
          type: string
        subcategory:
          type: string
        description:
          type: string
        price:
          type: number
        salePrice:
          type: number
          nullable: true
        images:
          type: array
          items:
            type: string
        rating:
          type: number
        reviewCount:
          type: integer
        stock:
          type: integer
        skinTypes:
          type: array
          items:
            type: string
        concerns:
          type: array
          items:
            type: string
        isBestSeller:
          type: boolean
        isNewArrival:
          type: boolean
        isFeatured:
          type: boolean
        ingredients:
          type: string
          nullable: true
        howToUse:
          type: string
          nullable: true
      required: [id, name, slug, category, subcategory, description, price, images, rating, reviewCount, stock, skinTypes, concerns, isBestSeller, isNewArrival, isFeatured]
    ProductDetail:
      allOf:
        - $ref: "#/components/schemas/Product"
        - type: object
          properties:
            reviews:
              type: array
              items:
                $ref: "#/components/schemas/Review"
    ProductsResponse:
      type: object
      properties:
        products:
          type: array
          items:
            $ref: "#/components/schemas/Product"
        total:
          type: integer
        page:
          type: integer
        totalPages:
          type: integer
      required: [products, total, page, totalPages]
...
    Review:
      type: object
      properties:
        id:
          type: integer
        productId:
          type: integer
        userId:
          type: integer
          nullable: true
        reviewerName:
          type: string
        rating:
          type: integer
        title:
          type: string
          nullable: true
        body:
          type: string
        skinType:
          type: string
          nullable: true
        source:
          type: string
          enum: [user, scraped]
        isApproved:
          type: boolean
        createdAt:
          type: string
      required: [id, productId, reviewerName, rating, body, source, isApproved, createdAt]
    CreateReviewBody:
      type: object
      properties:
        rating:
          type: integer
        title:
          type: string
          nullable: true
        body:
          type: string
        skinType:
          type: string
          nullable: true
      required: [rating, body]
    QuizAnswers:
      type: object
      properties:
        skinType:
          type: string
        concerns:
          type: array
          items:
            type: string
        ageRange:
          type: string
        dailyTime:
          type: string
        budget:
          type: string
      required: [skinType, concerns, ageRange, dailyTime, budget]
    QuizRecommendation:
      type: object
      properties:
        product:
          $ref: "#/components/schemas/Product"
        reason:
          type: string
      required: [product, reason]
    QuizResult:
      type: object
      properties:
        skinProfile:
          $ref: "#/components/schemas/QuizAnswers"
        recommendations:
          type: array
          items:
            $ref: "#/components/schemas/QuizRecommendation"
        createdAt:
          type: string
          nullable: true
      required: [skinProfile, recommendations]
...